Attempt at getting Adding and retrieving  users
Here I added java script files for index model and the controller.  In the model class I created two objects one for exercise app and one for the user. In the controller I attempted to use get and post to add to the users and exercises. There is a problem currently the server only receives a set of empty brackets and cannot post, current debugging